adjusted and splitted dispatcher
