+++ modem/arq_session.py
@@ -5,6 +5,8 @@ import structlog
 from event_manager import EventManager
 from modem_frametypes import FRAME_TYPE
 import time
+from arq_received_data_dispatcher import ARQReceivedDataDispatcher
+
 
 class ARQSession():
 
@@ -44,6 +46,7 @@ class ARQSession():
         self.frame_factory = data_frame_factory.DataFrameFactory(self.config)
         self.event_frame_received = threading.Event()
 
+        self.arq_received_data_dispatcher = ARQReceivedDataDispatcher()
         self.id = None
         self.session_started = time.time()
         self.session_ended = 0
@@ -88,10 +91,13 @@ class ARQSession():
         if self.state in self.STATE_TRANSITION:
             if frame_type in self.STATE_TRANSITION[self.state]:
                 action_name = self.STATE_TRANSITION[self.state][frame_type]
-                getattr(self, action_name)(frame)
+                received_data = getattr(self, action_name)(frame)
+                if received_data:
+                    self.arq_received_data_dispatcher.dispatch(received_data)
+
                 return
         
-        self.log(f"Ignoring unknow transition from state {self.state.name} with frame {frame['frame_type']}")
+        self.log(f"Ignoring unknown transition from state {self.state.name} with frame {frame['frame_type']}")
 
     def is_session_outdated(self):
         session_alivetime = time.time() - self.session_max_age
