+++ tests/test_data_dispatcher.py
@@ -2,21 +2,23 @@ import sys
 sys.path.append('modem')
 
 import unittest
-from data_dispatcher import DataDispatcher
+from arq_data_formatter import ARQDataFormatter
+from arq_received_data_dispatcher import ARQReceivedDataDispatcher
 
 class TestDispatcher(unittest.TestCase):
 
     @classmethod
     def setUpClass(cls):
-        cls.data_dispatcher = DataDispatcher()
+        cls.data_dispatcher = ARQReceivedDataDispatcher()
+        cls.data_formatter = ARQDataFormatter()
 
 
     def testEncapsulator(self):
         message_type = "p2pmsg"
         message_data = {"message": "Hello, P2P World!"}
 
-        encapsulated = self.data_dispatcher.encapsulate(message_data, message_type)
-        type, decapsulated = self.data_dispatcher.decapsulate(encapsulated.encode('utf-8'))
+        encapsulated = self.data_formatter.encapsulate(message_data, message_type)
+        type, decapsulated = self.data_formatter.decapsulate(encapsulated.encode('utf-8'))
         self.assertEqual(type, message_type)
         self.assertEqual(decapsulated, message_data)
 
@@ -24,7 +26,7 @@ class TestDispatcher(unittest.TestCase):
         message_type = "test"
         message_data = {"message": "Hello, P2P World!"}
 
-        encapsulated = self.data_dispatcher.encapsulate(message_data, message_type)
+        encapsulated = self.data_formatter.encapsulate(message_data, message_type)
         self.data_dispatcher.dispatch(encapsulated.encode('utf-8'))
